Extract shared money/month/gmail modules and fix four latent bugs
Acts on an architecture review of the hot-spot files (Dashboard.tsx,
server/routes.ts, History.tsx). Five findings, each closing a place where
one rule was restated in several files and the copies had drifted.

Session expiry (client/src/lib/queryClient.ts)
  The QueryCache onError handler keys off `message.startsWith("401:")`, but
  all 11 useQuery sites supplied their own queryFn, so the default was never
  reached; seven threw `${res.status}` — "401" with no colon. The condition
  could never be true and an expired session showed empty data instead of the
  login screen. Removed the redundant queryFns. useBudget keeps its own (404
  is the legitimate "no budget set" signal) but now shapes errors the same way.

shared/paise.ts — one owner for rupee <-> paise
  Conversion happened on a different tier depending on the entity: the client
  converted for expenses and budgets, the server for the other five. Both sent
  a bare `number`, so POST /api/expenses {amount:1500} meant Rs 15.00 while
  POST /api/income {amount:1500} meant Rs 1500.00. Standardised on integer
  paise across the wire (matching the DB and the Gmail endpoints), deleted the
  12 server-side conversions, and pinned amount fields to z.coerce.number().int()
  so a rupee decimal is a loud 400 rather than a silent 100x error.
  Fixes: commas in input ("1,234.56" parsed to 1), float drift (8.29 -> 828.999),
  NaN written to state when the Gmail amount field was cleared, and the budget
  being parsed twice.

shared/month.ts — one owner for every month figure
  Dashboard, Investments and widget-worker each restated the month rules and
  disagreed. The widget summed gross amounts, picked its top category on gross,
  and left SIP and EMI out of net cash flow entirely; Investments counted SIPs
  that had not started yet. On real May 2026 data the widget reported net cash
  flow of +Rs 2,245.55 when the truth was -Rs 8,054.45. All three now call
  monthSummary(). shared/emi.ts widened to a structural EmiLike so the worker
  can use it; the worker fetches rows instead of aggregates.
  Also: Investments and Emis captured the month key at module load, so a tab
  left open past midnight on the 1st kept reporting the old month.

shared/gmail.ts — one owner for the Gmail wire contract
  The staged-transaction shape was restated five times and the client copy had
  lost cardLast4, so the review UI could not correct a mis-routed card and tsc
  could not see the gap. Added a "Paid with" picker and a routing badge to the
  review step. Deleted server/gmail.ts (dead fifth copy).

Deleted ExpenseList.tsx and GmailSyncButton.tsx (284 unreachable lines)
  No importers, static or dynamic. They duplicated History's search, grouping
  and formatter but used gross amounts throughout, including a CSV export, so
  grepping for either found the wrong answer first. Gmail sync is unaffected —
  Dashboard mounts GmailSyncModal directly.

widget-worker credential
  The API key was accepted from ?key=, putting a live credential into request
  logs for an endpoint that returns the full month summary. Header-only now,
  with a constant-time compare; scriptable-widget.js updated to match.
  NOTE: rotate WIDGET_API_KEY, and deploy the worker and the phone script
  together or the widget will 401.

Verification — the repo has no test runner, so each module ships a guard that
runs on tsx with no new dependencies:
  script/verify-paise.ts       70 checks, incl. byte-for-byte parity with the
                               five formatter expressions it replaces
  script/verify-month.ts       26 checks, each divergence pinned with a case
                               the old behaviour would fail
  script/verify-query-seam.ts   5 checks, incl. a control proving the old
                               error shape does not trigger the handler
Also verified against the live routes and database (no writes): rupee decimals
now rejected on all five entities, staged cardLast4 round-trips, and both tiers
agree on every month figure. tsc clean for app and worker; build succeeds.

+++ shared/routes.ts
@@ -35,7 +35,7 @@ export const api = {
       method: 'POST' as const,
       path: '/api/expenses' as const,
       input: insertExpenseSchema.extend({
-        amount: z.coerce.number().positive(),
+        amount: z.coerce.number().int().positive(), // paise
       }),
       responses: {
         201: z.custom<typeof expenses.$inferSelect>(),
@@ -46,7 +46,7 @@ export const api = {
       method: 'PUT' as const,
       path: '/api/expenses/:id' as const,
       input: insertExpenseSchema.partial().extend({
-        amount: z.coerce.number().positive().optional(),
+        amount: z.coerce.number().int().positive().optional(), // paise
       }),
       responses: {
         200: z.custom<typeof expenses.$inferSelect>(),
@@ -77,7 +77,7 @@ export const api = {
       path: '/api/budgets' as const,
       input: z.object({
         month: z.string(),
-        amount: z.coerce.number().positive(),
+        amount: z.coerce.number().int().positive(), // paise
       }),
       responses: {
         200: z.custom<any>(),
